review comment: move error logic to different method
